Improved UI design
Changes to Level Select:
- Added placeholder background images
- Added pulsing color changes
- Added background music
       plays throughout the whole game continuously
- Added click sfx for toGame buttons
- Added fonts for level select panels
- Made general aesthetic changes
- Made level select panels into prefabs for ease of editing

Changes to Game scene:
- Added keyboard click for WASD
- Fixed timer bug that would round up to 1 minute at the 30 second mark

Things to do:
- Transfer game board to a UI element or grid layout
- Add star functionality
- Add information about password length to SetMap script
- Improve UI for other scenes
- SFX for all buttons, SFX for incorrect key.

+++ Comfy Knitting/Assets/Scripts/DisplayData.cs
@@ -22,8 +22,8 @@ public class DisplayData : MonoBehaviour
         return "N/A";
       }
       else{
-        float seconds = x % 60;
-        float minutes = x / 60;
+        int seconds = (int)(x % 60);
+        int minutes = (int)(x / 60);
         return minutes.ToString("00") + ":" + seconds.ToString("00");
       }
     }
